+++ ZoomChartsTimeChartFree/src/visual.ts
@@ -9,6 +9,8 @@ module powerbi.extensibility.visual {
         protected dataSourceIdentity: string = "";
         protected pendingSettings: ZoomCharts.Configuration.TimeChartSettings = {};
         protected updateTimer: number;
+        protected primaryAxisFormatString: string = "#,0.00";
+        protected secondaryAxisFormatString: string = "#,0.00";
         protected lastTimeRange: [number, number] = [null, null];
         protected colors: IColorPalette;
         protected selectionManager: ISelectionManager;
@@ -25,8 +27,8 @@ module powerbi.extensibility.visual {
         public currentScale: any = 1;
 
         constructor(options: VisualConstructorOptions) {
-            version = "v1.1.1.2";
-            releaseDate = "Mar 5, 2019";
+            version = "v1.2.0.0";
+            releaseDate = "Mar 6, 2019";
             visualType = "advanced-time-chart";
             visualName = "Advanced Timeline Chart Visual";
             this.target = options.element;
@@ -231,6 +233,8 @@ module powerbi.extensibility.visual {
                     return x;
                 });
             }
+
+            setAxisFormatting(series, this);
 
             this.series = series;