+++ ZoomChartsFacetChartFree/release-notes.md
@@ -2,6 +2,16 @@
 
 =========================== Changelog ===========================
 
+Version 1.2.0.0 (2019-03-06)
+
+* Issue Number: BS-128.
+  Added formatting default strings for primary and secondary axis.
+  Added function to correctly set formatting string for primary 
+  and secondary axis.
+* Added correct versioning start point.
+
+-----------------------------------------------------------------
+
 Version 1.1.1.2 (2019-03-05)
 
 * Issue Number: BS-143.